SFA: Start implementing heat shimmer (WIP)
